+++ libaegisub/include/libaegisub/thesaurus.h
@@ -31,7 +31,7 @@ namespace charset { class IconvWrapper; }
 
 class Thesaurus {
 	/// Map of word -> byte position in the data file
-	boost::container::flat_map<std::string, int> offsets;
+	boost::container::flat_map<std::string, size_t> offsets;
 	/// Read handle to the data file
 	std::unique_ptr<read_file_mapping> dat;
 	/// Converter from the data file's charset to UTF-8
